Better game camera positioning, added all needed sounds and small code changes.

+++ src/main/java/org/example/utils/Settings.java
@@ -4,6 +4,6 @@ public class Settings {
     public static float GRID_WIDTH = 0.25f;
     public static float  GRID_HEIGHT = 0.25f;
 
-    public static float  SCREEN_WIDTH = 1920;
-    public static float  SCREEN_HEIGHT = 1080;
+    public static int  SCREEN_WIDTH = 1920;
+    public static int  SCREEN_HEIGHT = 1080;
 }
